Change terminology for floating-point types from `Decimal` to `Real`
Phew, we can all stop wondering whether decimal means "floating-point"
or "base-10 integer".

+++ delphi-frontend/src/main/java/au/com/integradev/delphi/preprocessor/directive/expression/Expression.java
@@ -33,7 +33,7 @@ public interface Expression {
     UNKNOWN,
     STRING,
     INTEGER,
-    DECIMAL,
+    REAL,
     BOOLEAN,
     SET
   }
@@ -59,7 +59,7 @@ public interface Expression {
       return BigInteger.ZERO;
     }
 
-    default Double asDecimal() {
+    default Double asDouble() {
       return 0.0;
     }
